refactor hooks and handlers #26
