+++ src/components/Canvas/handlers/index.ts
@@ -1,0 +1,5 @@
+export { default as useCustomizationHandler } from './useCustomizationHandler'
+export { default as useEventsHandler } from './useEventsHandler'
+export { default as useZoomHandler } from './useZoomHandler'
+export { default as useCoreHandler } from './useCoreHandler'
+export { default as useContainerHandler } from './useContainerHandler'
