+++ src/components/Canvas/hooks/index.ts
@@ -1,6 +1,1 @@
 export { default as useCanvasContext } from './useCanvasContext'
-export { default as useFabricSettings } from './useFabricSettings'
-export { default as useEventsHandler } from './useEventsHandler'
-export { default as useCoreHandler } from './useCoreHandler'
-export { default as useContainerHandler } from './useContainerHandler'
-export { default as useZoomHandler } from './useZoomHandler'
