+++ src/components/Canvas/hooks/useCanvasContext.ts
@@ -7,9 +7,7 @@ function useCanvasContext() {
     zoomRatio,
     setZoomRatio,
     setCanvas,
-
     canvas,
-
     activeObject,
     setActiveObject,
   }
